requirements.txt
pandas>=2.0.0
pyyaml>=6.0
anthropic>=0.40.0
jinja2>=3.1.0
click>=8.0.0
email-validator>=2.0.0
pytest>=7.0.0
